add frontend login, logout, authentication, styling
